Add show into dataset API
